+++ src/Symfony/Component/OptionsResolver/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 -----
 
  * added fluent configuration of options using `OptionResolver::define()`
+ * added `setInfo()` and `getInfo()` methods
 
 5.0.0
 -----